requirements.txt
streamlit>=1.28.0
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
Pillow>=9.0.0
pandas>=1.5.0
fpdf2>=2.7.0
numpy>=1.24.0